+++ packages/store/tests/build-action-handler/action-director.spec.ts
@@ -1,0 +1,31 @@
+import { TestBed } from '@angular/core/testing';
+import { ActionDirector, provideStore, Store } from '@ngxs/store';
+
+import { AddCountry } from './fixtures/actions';
+import { CountriesState } from './fixtures/state';
+
+describe('ActionDirector', () => {
+  it('should allow attaching action handlers on the fly', async () => {
+    // Arrange
+    TestBed.configureTestingModule({
+      providers: [provideStore([CountriesState])]
+    });
+
+    const store = TestBed.inject(Store);
+
+    // Act
+    store.dispatch(new AddCountry('Canada'));
+
+    // Assert
+    expect(store.snapshot()).toEqual({ countries: [] });
+
+    // Act
+    const { addCountryActionHandler } = await import('./fixtures/add-country-action-handler');
+    addCountryActionHandler(TestBed.inject(ActionDirector));
+
+    store.dispatch(new AddCountry('Canada'));
+
+    // Assert
+    expect(store.snapshot()).toEqual({ countries: ['Canada'] });
+  });
+});